Use provider HOC instead of 'next-redux-wrapper'
