+++ package.json
@@ -21,7 +21,6 @@
     "immutable": "^3.8.2",
     "lodash": "^4.17.4",
     "next": "^4.2.1",
-    "next-redux-wrapper": "^1.3.5",
     "normalizr": "^3.2.4",
     "path-to-regexp": "^2.1.0",
     "react": "^16.2.0",
